Extract 23137 repro
